LayoutTests/fast/events/wheel/wheel-event-listeners-on-body-made-passive-expected.txt

CONSOLE MESSAGE: Unable to preventDefault() on event of type 'wheel': Inside passive event listener.
CONSOLE MESSAGE: Unable to preventDefault() on event of type 'wheel': Inside passive event listener.
Tests that a wheel event handler on the body is passive (cannot prevent scrolling)
PASS windowScrollEventCount > 0 is true
PASS defaultPrevented is false
PASS successfullyParsed is true

TEST COMPLETE
